solving equations and spelling typo fixes

+++ include/bit_operations.h
@@ -10,13 +10,13 @@
 #include <stdint.h>
 #include "calc_status.h"
 
-calc_status_t l_shift(uint32_t f_num, uint32_t s_num, uint32_t * p_result);
-calc_status_t r_shift(uint32_t f_num, uint32_t s_num, uint32_t * p_result);
-calc_status_t op_and(uint32_t f_num, uint32_t s_num, uint32_t * p_result);
-calc_status_t op_or(uint32_t f_num, uint32_t s_num, uint32_t * p_result);
-calc_status_t op_xor(uint32_t f_num, uint32_t s_num, uint32_t * p_result);
-calc_status_t l_rotate(uint32_t f_num, uint32_t s_num, uint32_t * p_result);
-calc_status_t r_rotate(uint32_t f_num, uint32_t s_num, uint32_t * p_result);
+calc_status_t l_shift(uint64_t f_num, uint64_t s_num, uint64_t * p_result);
+calc_status_t r_shift(uint64_t f_num, uint64_t s_num, uint64_t * p_result);
+calc_status_t op_and(uint64_t f_num, uint64_t s_num, uint64_t * p_result);
+calc_status_t op_or(uint64_t f_num, uint64_t s_num, uint64_t * p_result);
+calc_status_t op_xor(uint64_t f_num, uint64_t s_num, uint64_t * p_result);
+calc_status_t l_rotate(uint64_t f_num, uint64_t s_num, uint64_t * p_result);
+calc_status_t r_rotate(uint64_t f_num, uint64_t s_num, uint64_t * p_result);
 
 #endif
