+++ include/calc_operations.h
@@ -10,11 +10,11 @@
 #include <stdint.h>
 #include "calc_status.h"
 
-calc_status_t add(int32_t f_num, int32_t s_num, int32_t * result);
-calc_status_t subtract(int32_t f_num, int32_t s_num, int32_t * result);
-calc_status_t multiply(int32_t f_num, int32_t s_num, int32_t * result);
-calc_status_t divide(int32_t f_num, int32_t s_num, int32_t * result);
-calc_status_t modulo(int32_t f_num, int32_t s_num, int32_t * result);
+calc_status_t add(int64_t f_num, int64_t s_num, int64_t * result);
+calc_status_t subtract(int64_t f_num, int64_t s_num, int64_t * result);
+calc_status_t multiply(int64_t f_num, int64_t s_num, int64_t * result);
+calc_status_t divide(int64_t f_num, int64_t s_num, int64_t * result);
+calc_status_t modulo(int64_t f_num, int64_t s_num, int64_t * result);
 
 #endif
