+++ src/calc_operations.c
@@ -8,7 +8,7 @@
 #include "calc_status.h"
 #include "calc_operations.h"
 
-calc_status_t add(int32_t f_num, int32_t s_num, int32_t * p_result)
+calc_status_t add(int64_t f_num, int64_t s_num, int64_t * p_result)
 {
 	calc_status_t status = CALC_STATUS_OK;
 
@@ -16,8 +16,8 @@ calc_status_t add(int32_t f_num, int32_t s_num, int32_t * p_result)
 	{
 		status = CALC_STATUS_NULL_POINTER;
 	}
-	else if (((0 < s_num) && ((INT32_MAX - s_num) < f_num)) ||
-			 ((0 > s_num) && ((INT32_MIN -s_num) > f_num)))
+	else if (((0 < s_num) && ((INT64_MAX - s_num) < f_num)) ||
+			 ((0 > s_num) && ((INT64_MIN -s_num) > f_num)))
 	{
 		status = CALC_STATUS_OVERFLOW;
 	}
@@ -29,7 +29,7 @@ calc_status_t add(int32_t f_num, int32_t s_num, int32_t * p_result)
 	return status;
 }
 
-calc_status_t subtract(int32_t f_num, int32_t s_num, int32_t * p_result)
+calc_status_t subtract(int64_t f_num, int64_t s_num, int64_t * p_result)
 {
 	calc_status_t status = CALC_STATUS_OK;
 
@@ -37,8 +37,8 @@ calc_status_t subtract(int32_t f_num, int32_t s_num, int32_t * p_result)
 	{
 		status = CALC_STATUS_NULL_POINTER;
 	}
-	else if (((0 < s_num) && ((INT32_MIN + s_num) > f_num)) ||
-		   	 ((0 > s_num) && ((INT32_MAX + s_num) < f_num)))
+	else if (((0 < s_num) && ((INT64_MIN + s_num) > f_num)) ||
+		   	 ((0 > s_num) && ((INT64_MAX + s_num) < f_num)))
 	{
 		status = CALC_STATUS_OVERFLOW;
 	}
@@ -50,7 +50,7 @@ calc_status_t subtract(int32_t f_num, int32_t s_num, int32_t * p_result)
 	return status;
 }
 
-calc_status_t multiply(int32_t f_num, int32_t s_num, int32_t * p_result)
+calc_status_t multiply(int64_t f_num, int64_t s_num, int64_t * p_result)
 {
 	calc_status_t status = CALC_STATUS_OK;
 
@@ -58,12 +58,12 @@ calc_status_t multiply(int32_t f_num, int32_t s_num, int32_t * p_result)
 	{
 		status = CALC_STATUS_NULL_POINTER;
 	}
-	else if (((-1 == f_num) && (INT32_MIN == s_num)) ||
-			 ((-1 == s_num) && (INT32_MIN == f_num)) ||
-			 ((0 < f_num) && (0 < s_num) && ((INT32_MAX / s_num) < f_num)) ||
-			 ((0 < f_num) && (0 > s_num) && ((INT32_MAX / f_num) > s_num)) ||
-			 ((0 > f_num) && (0 < s_num) && ((INT32_MAX / s_num) > f_num)) ||
-			 ((0 > f_num) && (0 > s_num) && ((INT32_MAX / s_num) > f_num)))
+	else if (((-1 == f_num) && (INT64_MIN == s_num)) ||
+			 ((-1 == s_num) && (INT64_MIN == f_num)) ||
+			 ((0 < f_num) && (0 < s_num) && ((INT64_MAX / s_num) < f_num)) ||
+			 ((0 < f_num) && (0 > s_num) && ((INT64_MAX / f_num) > s_num)) ||
+			 ((0 > f_num) && (0 < s_num) && ((INT64_MAX / s_num) > f_num)) ||
+			 ((0 > f_num) && (0 > s_num) && ((INT64_MAX / s_num) > f_num)))
 	{
 		status = CALC_STATUS_OVERFLOW;
 	}
@@ -75,7 +75,7 @@ calc_status_t multiply(int32_t f_num, int32_t s_num, int32_t * p_result)
 	return status;
 }
 
-calc_status_t divide(int32_t f_num, int32_t s_num, int32_t *p_result)
+calc_status_t divide(int64_t f_num, int64_t s_num, int64_t *p_result)
 {
 	calc_status_t status = CALC_STATUS_OK;
 
@@ -87,7 +87,7 @@ calc_status_t divide(int32_t f_num, int32_t s_num, int32_t *p_result)
 		printf("Error: can not divide by 0\n");
 		status = CALC_STATUS_DIVIDE_BY_ZERO;
 	}
-	else if ((INT32_MIN == f_num) && (-1 == s_num)){
+	else if ((INT64_MIN == f_num) && (-1 == s_num)){
 		status = CALC_STATUS_OVERFLOW;
 	}
 	else
@@ -98,7 +98,7 @@ calc_status_t divide(int32_t f_num, int32_t s_num, int32_t *p_result)
 	return status;
 }
 
-calc_status_t modulo(int32_t f_num, int32_t s_num, int32_t *p_result)
+calc_status_t modulo(int64_t f_num, int64_t s_num, int64_t *p_result)
 {
 	calc_status_t status = CALC_STATUS_OK;
 
@@ -110,7 +110,7 @@ calc_status_t modulo(int32_t f_num, int32_t s_num, int32_t *p_result)
 		printf("Error: can not divide by 0\n");
 		status = CALC_STATUS_DIVIDE_BY_ZERO;
 	}
-	else if ((INT32_MIN == f_num) && (-1 == s_num)){
+	else if ((INT64_MIN == f_num) && (-1 == s_num)){
 		status = CALC_STATUS_OVERFLOW;
 	}
 	else
